requirements.txt
asgiref==3.2.7
astroid==2.4.2
autopep8==1.5.3
click==7.1.2
dj-database-url==0.5.0
Django==3.0.7
django-extensions==3.0.1
django-filter==2.3.0
django-rest-framework==0.1.0
django-rq==2.3.2
djangorestframework==3.11.0
efficientnet-pytorch==0.6.3
environs==8.0.0
furl==2.1.0
future==0.18.2
gunicorn==20.0.4
honcho==1.0.1
imutils==0.5.3
isort==4.3.21
lazy-object-proxy==1.4.3
marshmallow==3.6.1
mccabe==0.6.1
munch==2.5.0
numpy==1.18.5
opencv-python==4.2.0.34
orderedmultidict==1.0.1
Pillow==7.1.2
pretrainedmodels==0.7.4
psycopg2==2.8.5
pycodestyle==2.6.0
pylint==2.5.3
python-dotenv==0.13.0
pytz==2020.1
redis==3.5.3
rq==1.4.2
scipy==1.4.1
segmentation-models-pytorch==0.1.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
https://download.pytorch.org/whl/cpu/torch-1.3.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.4.2%2Bcpu-cp36-cp36m-linux_x86_64.whl
tqdm==4.46.1
typed-ast==1.4.1
wrapt==1.12.1
